+++ apps/examples/src/main/java/examples/features/async/AsyncExampleFIFO.java
@@ -33,7 +33,7 @@ public class AsyncExampleFIFO extends Node {
 
     public static void main(String[] args) {
         Topology tp = new Topology();
-        tp.setClockSpeed(100);
+        tp.setTimeUnit(100);
         tp.setMessageEngine(new AsyncMessageEngine(20, AsyncMessageEngine.Type.FIFO));
         //tp.getMessageEngine().setSpeed(10);
         tp.setDefaultNodeModel(AsyncExampleFIFO.class);